refactor address routing
