+++ src/presentation/controller/addressLookUp/routing/limitedPartnership.ts
@@ -1,0 +1,126 @@
+import AddressPageType from "../PageType";
+import { TERM_URL, WHERE_IS_THE_JURISDICTION_URL } from "../../registration/url";
+import * as url from "../url";
+
+export enum AddressCacheKeys {
+  addressCacheKey = "addressCacheKey",
+  territoryCacheKey = "territoryCacheKey"
+}
+
+// Registered Office Address
+
+const registeredOfficeAddressCacheKeys = {
+  [AddressCacheKeys.addressCacheKey]: "registered_office_address"
+};
+
+const addressRoutingPostcodeRegisteredOfficeAddress = {
+  previousUrl: WHERE_IS_THE_JURISDICTION_URL,
+  currentUrl: url.POSTCODE_REGISTERED_OFFICE_ADDRESS_URL,
+  nextUrl: url.CHOOSE_REGISTERED_OFFICE_ADDRESS_URL,
+  pageType: AddressPageType.postcodeRegisteredOfficeAddress,
+  data: {
+    ...registeredOfficeAddressCacheKeys,
+    confirmAddressUrl: url.CONFIRM_REGISTERED_OFFICE_ADDRESS_URL,
+    enterManualAddressPageType: AddressPageType.enterRegisteredOfficeAddress
+  }
+};
+
+const addressRoutingChooseRegisteredOfficeAddress = {
+  previousUrl: url.POSTCODE_REGISTERED_OFFICE_ADDRESS_URL,
+  currentUrl: url.CHOOSE_REGISTERED_OFFICE_ADDRESS_URL,
+  nextUrl: url.CONFIRM_REGISTERED_OFFICE_ADDRESS_URL,
+  pageType: AddressPageType.chooseRegisteredOfficeAddress,
+  data: {
+    ...registeredOfficeAddressCacheKeys,
+    enterManualAddressPageType: AddressPageType.enterRegisteredOfficeAddress
+  }
+};
+
+const addressRoutingEnterRegisteredOfficeAddress = {
+  previousUrl: url.POSTCODE_REGISTERED_OFFICE_ADDRESS_URL,
+  currentUrl: url.ENTER_REGISTERED_OFFICE_ADDRESS_URL,
+  nextUrl: url.CONFIRM_REGISTERED_OFFICE_ADDRESS_URL,
+  pageType: AddressPageType.enterRegisteredOfficeAddress,
+  data: {
+    ...registeredOfficeAddressCacheKeys
+  }
+};
+
+const addressRoutingConfirmRegisteredOfficeAddress = {
+  previousUrl: url.POSTCODE_REGISTERED_OFFICE_ADDRESS_URL,
+  currentUrl: url.CONFIRM_REGISTERED_OFFICE_ADDRESS_URL,
+  nextUrl: url.POSTCODE_PRINCIPAL_PLACE_OF_BUSINESS_ADDRESS_URL,
+  pageType: AddressPageType.confirmRegisteredOfficeAddress,
+  data: {
+    ...registeredOfficeAddressCacheKeys,
+    enterManualAddressPageType: AddressPageType.enterRegisteredOfficeAddress
+  }
+};
+
+const registeredOfficeAddress = [
+  addressRoutingPostcodeRegisteredOfficeAddress,
+  addressRoutingChooseRegisteredOfficeAddress,
+  addressRoutingEnterRegisteredOfficeAddress,
+  addressRoutingConfirmRegisteredOfficeAddress
+];
+
+// principal place of business
+
+const principalPlaceOfBusinessAddressCacheKeys = {
+  [AddressCacheKeys.addressCacheKey]: "principal_place_of_business_address"
+};
+
+const addressRoutingPostcodePrincipalPlaceOfBusinessAddress = {
+  previousUrl: url.CONFIRM_REGISTERED_OFFICE_ADDRESS_URL,
+  currentUrl: url.POSTCODE_PRINCIPAL_PLACE_OF_BUSINESS_ADDRESS_URL,
+  nextUrl: url.CHOOSE_PRINCIPAL_PLACE_OF_BUSINESS_ADDRESS_URL,
+  pageType: AddressPageType.postcodePrincipalPlaceOfBusinessAddress,
+  data: {
+    ...principalPlaceOfBusinessAddressCacheKeys,
+    enterManualAddressPageType: AddressPageType.enterPrincipalPlaceOfBusinessAddress,
+    confirmAddressUrl: url.CONFIRM_PRINCIPAL_PLACE_OF_BUSINESS_ADDRESS_URL
+  }
+};
+
+const addressRoutingChoosePrincipalPlaceOfBusinessAddress = {
+  previousUrl: url.POSTCODE_PRINCIPAL_PLACE_OF_BUSINESS_ADDRESS_URL,
+  currentUrl: url.CHOOSE_PRINCIPAL_PLACE_OF_BUSINESS_ADDRESS_URL,
+  nextUrl: url.CONFIRM_PRINCIPAL_PLACE_OF_BUSINESS_ADDRESS_URL,
+  pageType: AddressPageType.choosePrincipalPlaceOfBusinessAddress,
+  data: {
+    ...principalPlaceOfBusinessAddressCacheKeys,
+    enterManualAddressPageType: AddressPageType.enterPrincipalPlaceOfBusinessAddress
+  }
+};
+
+const addressRoutingEnterPrincipalPlaceOfBusinessAddress = {
+  previousUrl: url.POSTCODE_PRINCIPAL_PLACE_OF_BUSINESS_ADDRESS_URL,
+  currentUrl: url.ENTER_PRINCIPAL_PLACE_OF_BUSINESS_ADDRESS_URL,
+  nextUrl: url.CONFIRM_PRINCIPAL_PLACE_OF_BUSINESS_ADDRESS_URL,
+  pageType: AddressPageType.enterPrincipalPlaceOfBusinessAddress,
+  data: {
+    ...principalPlaceOfBusinessAddressCacheKeys
+  }
+};
+
+const addressRoutingConfirmPrincipalPlaceOfBusinessAddress = {
+  previousUrl: url.POSTCODE_PRINCIPAL_PLACE_OF_BUSINESS_ADDRESS_URL,
+  currentUrl: url.CONFIRM_PRINCIPAL_PLACE_OF_BUSINESS_ADDRESS_URL,
+  nextUrl: TERM_URL,
+  pageType: AddressPageType.confirmPrincipalPlaceOfBusinessAddress,
+  data: {
+    ...principalPlaceOfBusinessAddressCacheKeys,
+    enterManualAddressPageType: AddressPageType.enterPrincipalPlaceOfBusinessAddress
+  }
+};
+
+const principalPlaceOfBusinessAddress = [
+  addressRoutingPostcodePrincipalPlaceOfBusinessAddress,
+  addressRoutingChoosePrincipalPlaceOfBusinessAddress,
+  addressRoutingEnterPrincipalPlaceOfBusinessAddress,
+  addressRoutingConfirmPrincipalPlaceOfBusinessAddress
+];
+
+const limitedPartnershipRouting = [...registeredOfficeAddress, ...principalPlaceOfBusinessAddress];
+
+export default limitedPartnershipRouting;
